fix: Fixed Pie chart labels and colors for all charts.
