+++ budget-app/src/helpers/helpers.js
@@ -37,19 +37,45 @@ export const getMinDate = (accountId, accounts) => {
   return correctAccount[0].created_at
 }
 
-export const dataColors = [
-  '#8DB55C',
-  '#498F60',
-  '#E2E062',
-  '#297159',
-  '#E18DB7',
-  '#ADA0D3',
-  '#69B0D2',
-  '#73AC61',
-  '#15534C',
-  '#3ABAB4',
-  '#A6C760',
-  '#58BB87',
+export const getExpensesForPieChart = transactions => {
+  // Only expenses
+  const expenses = transactions
+    .filter(transaction => transaction.type === 'expense')
+    .map(expense => {
+      return { ...expense, amount: expense.amount * -1 }
+    })
+
+  // 1. Get a list of all categories.
+  const categoriesAll = expenses.map(expense => expense.category)
+  const catUniques = [...new Set(categoriesAll)]
+
+  // 2. loop over the categories and return only the expenses that belong to that category.
+
+  return catUniques.reduce((acc, prev) => {
+    // Get the sum of each category
+    const sum = expenses.reduce((total, amount) => {
+      // Check if the category is the same in the expense as the one we loop through. If it is, then add the amount to total
+      if (prev === amount.category) {
+        return Number(total) + Number(amount.amount)
+      }
+      return total
+    }, 0)
+    // 3. Sum up the amounts for each category.
+    return [...acc, { name: prev, value: sum }]
+  }, [])
+}
+
+export const COLORS = [
+  '#a6cee3',
+  '#1f78b4',
+  '#b2df8a',
+  '#33a02c',
+  '#fb9a99',
+  '#e31a1c',
+  '#fdbf6f',
+  '#ff7f00',
+  '#cab2d6',
+  '#6a3d9a',
 ]
 export const locales = [
   'AED',
